+++ src/StateSmith/Runner/SmRunner.cs
@@ -123,6 +123,7 @@ public class SmRunner : SmRunner.IExperimentalAccess
             services.AddSingleton(new ExpansionConfigReaderObjectProvider(iRenderConfig));
             services.AddSingleton(settings); // todo_low - split settings up more
             services.AddSingleton<ExpansionsPrep>();
+            services.AddSingleton<FilePathPrinter>(new FilePathPrinter(settings.filePathPrintBase.ThrowIfNull()));
         });
 
         AlgoOrTranspilerUpdated();